requirements.txt
altgraph==0.17.5
annotated-types==0.7.0   
anyio==4.12.1
beautifulsoup4==4.14.3   
certifi==2025.11.12      
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
EasyProcess==1.1
elevate==0.1.3
entrypoint2==1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
packaging==25.0
patool==4.0.2
pefile==2023.2.7
pillow==12.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.10
PyMySQL==1.1.2
pyunpack==0.3
pywin32-ctypes==0.2.3
requests==2.32.5
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.8.1
tqdm==4.67.1
translatekit==0.2.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0